trading-app/requirements.txt
ib_insync>=0.9.70
pandas
numpy
psutil>=5.9.0
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.7.0
pymongo==4.6.2
